passage en temps réel

+++ fonctions.h
@@ -5,10 +5,10 @@ int x_absolute(int x_centre, int x_relatif);
 int y_absolute(int x_centre, int x_relatif);
 
 /* Renvoie la coordonée x d'un point sur un cercle de rayon 1 en fonction du temps avec 1 sec = 1 jour*/
-float calculAbscisse(float secondes);
+float calculAbscisse(double secondes);
 
 /* Renvoie la coordonée x d'un point sur un cercle de rayon 1 en fonction du temps avec 1 sec = 1 jour*/
-float calculOrdonee(float secondes);
+float calculOrdonee(double secondes);
 
 /* Affiche le nom d'un objet au dessus de celui-ci */
 void centre_text(int x_objet, int y_objet, int rayon_objet, char nom_objet[]);
